tourism-ai-platform/ai-services/orchestrator-svc/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
tenacity==8.2.3
aiokafka==0.10.0
boto3==1.34.45
langgraph==0.0.21
langsmith==0.1.29
redis==5.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
